// Catalog/graphql/Catalog.generated.tsx
export type Maybe<T> = T | null;
// ...
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  BigDecimal: any;
  Date: any;
  DateTime: any;
  GqlDuration: any;
  JwtPermissions: any;
  Permissions: any;
  Uuid: any;
};
// ...
export type EquipmentClassType = {
  __typename?: 'EquipmentClassType';
  /** Комментарий */
  comment: Scalars['String'];
  /** Дата создания записи */
  createdAt: Scalars['DateTime'];
  /** Ключ к картинке по умолчанию для моделей данного класса техники, доступно через сервис file-storage */
  defaultModelImage?: Maybe<Scalars['Uuid']>;
  /** Идентификатор класса техники */
  id: Scalars['ID'];
  /** Модели принадлежащие данному классу техники */
  models: EquipmentModelTypeConnection;
  /** Массив ключей фотографий, фото хранятся в файловом хранилище, доступ через сервис file-storage */
  photos: Array<Scalars['Uuid']>;
  /** Уникальный токен */
  slug: Scalars['String'];
  /** Описание */
  title: Scalars['String'];
  /** Тип класса техники, используется на фронте для сортировки */
  type: Scalars['String'];
  /** Дата последнего обновления записи */
  updatedAt: Scalars['DateTime'];
};
// ...
export type EquipmentClassTypeEdge = {
  __typename?: 'EquipmentClassTypeEdge';
  /** cursor marks a unique position or index into the connection */
  cursor: Scalars['String'];
  /** The item at the end of the edge */
  node?: Maybe<EquipmentClassType>;
};
// ...
export type EquipmentModelTypeConnection = {
  __typename?: 'EquipmentModelTypeConnection';
  /** a list of edges */
  edges?: Maybe<Array<Maybe<EquipmentModelTypeEdge>>>;
  /** details about this specific page */
  pageInfo: PageInfo;
};